requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1745255672738/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1734342155601/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1740094013202/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725571112467/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1740893557677/work
frozenlist @ file:///home/conda/feedstock_root/build_artifacts/frozenlist_1737645236190/work
google-api-core @ file:///home/conda/feedstock_root/build_artifacts/google-api-core-split_1741643016905/work
google-api-python-client @ file:///home/conda/feedstock_root/build_artifacts/google-api-python-client_1746017233360/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1744688154288/work
google-auth-httplib2 @ file:///home/conda/feedstock_root/build_artifacts/google-auth-httplib2_1733907410727/work
google-auth-oauthlib==1.2.2
googleapis-common-protos @ file:///home/conda/feedstock_root/build_artifacts/googleapis-common-protos-feedstock_1744688777945/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httplib2 @ file:///home/conda/feedstock_root/build_artifacts/httplib2_1733927481809/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1744972419893/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1733752848439/work
propcache @ file:///home/conda/feedstock_root/build_artifacts/propcache_1744524932661/work
proto-plus @ file:///home/conda/feedstock_root/build_artifacts/proto-plus_1741676149446/work
protobuf @ file:///home/conda/feedstock_root/build_artifacts/protobuf_1741124571372/work/bazel-bin/python/dist/protobuf-5.29.3-cp39-abi3-linux_x86_64.whl#sha256=7a950517418fb10c23170e0132324ba319529b6002f25fbd96534cd9bf72d4d3
pyasn1 @ file:///home/conda/feedstock_root/build_artifacts/pyasn1_1733217608156/work
pyasn1_modules @ file:///home/conda/feedstock_root/build_artifacts/pyasn1-modules_1743436035994/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1732782409051/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1737243356468/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1733738580568/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1733772243268/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1744825426113/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
uritemplate @ file:///home/conda/feedstock_root/build_artifacts/uritemplate_1733927878495/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1744972437465/work
zstandard==0.23.0
